fix: coordinates are doubles!

+++ database/migrations/2025_02_02_182805_create_map_markers_table.php
@@ -18,8 +18,8 @@ return new class extends Migration {
             $table->string('name');
             $table->string('description');
             $table->string('address');
-            $table->integer('lat');
-            $table->integer('lon');
+            $table->double('lat');
+            $table->double('lon');
             $table->timestamps();
         });
     }
